requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.8
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
Faker==12.0.1
idna==3.4
Markdown==3.4.1
mixer==7.2.2
mock==5.0.1
projects==1.0.9
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
urllib3==1.26.15
urwid==2.1.2
